passato ai fragment tutto ai fragment
create tutte 5 i fragment per la bottomview, beelist, home, map, social, userarea

+++ app/src/main/java/com/dustolab/beerapp/model/Bar.kt
@@ -7,7 +7,7 @@ data class Bar(
     var beerList: List<BeerInBar>? = null,
     var drink: List<MenuEntry>? = null,
     var food: List<MenuEntry>? = null,
-    var rating: Double? = null,
+    var rating: Float? = null,
     var favoriteBy: List<String>? = null
 ): Record(){
